fix: logoValidator funciton and minor fixes
